Reduced number of operations in block sieve code
Changed max(start_idx, 2) * p to max(start_idx, p) * p. This is done so that we always start from the max. of the smallest multiple of p in the range [l, r] and p * p. It is more beneficial to start from p * p rather than 2 * p as lower multiples of p would have been covered by the smaller prime factors.