Added more information in finding negative cycle using Bellman-Ford
The information is taken from the [Bellman-Ford](https://cp-algorithms.com/graph/bellman_ford.html) page. I felt it should be present here as well to make this article self-sufficient.

Also modified the code to remove the redundant check since `d` is initialised to `0` instead of `INF` so the check is not required.