Update rmq.md
Fenwick tree can be build in O(N). Instead of calling update on each entry, iterate through each element and update the array's value to the immediate parent. And we can use Fenwick's tree for L != 1 also by computing R and inverse function with computed L. i.e. if summation is the function then you find for R and L and subtract the results of L from R to get the answer.
Link: https://stackoverflow.com/a/66027174
Check comments on the answer