Update bellman_ford.md
The claim that the vertex y would be the first vertex reachable from the source is misleading. A counter example is given below:
n=5 m=5
edges vector:
a:1 b:2 w:1
a:2 b:3 w:2
a:3 b:4 w:2
a:4 b:3 w:(-4)
a:4 b:5 w:1


Clearly in the 5th iteration the last relaxation would be on the vertex number 5, now looping back to the predecessor 5 times would give us the vertex number 4 i.e. y=4. However the first vertex reachable fromt the source and part of the negative cycle is vertex number 3. Thus this claim is misleading.