fix: draft: invariant loop condition
The proof states that a transition point exists, i.e f(0)=0, f(n-1) = 1 . The proof also states that the loop maintains the invariant f(l)=0 and f(r) = 1 when r-l >1 . Strict loop invariance assumes that the condition holds before, during, and after the loop. for l=-1 f(l) is not 0, its undefined.
We can easily bypass that by starting from zero. in this implementation we dont need edge-case work, as we have a guaranteed transition point.